+++ src/main/java/com/mageddo/dnsproxyserver/server/dns/solver/docker/Container.java
@@ -40,6 +40,7 @@ public class Container {
   }
 
   public IP geDefaultIp(IP.Version version, String networkName) {
+    // fixme #444 must implement it
     throw new UnsupportedOperationException();
   }
