+++ src/main/java/com/mageddo/dnsproxyserver/server/dns/solver/docker/dataprovider/DockerDAODefault.java
@@ -4,7 +4,7 @@ import com.github.dockerjava.api.DockerClient;
 import com.mageddo.commons.lang.Objects;
 import com.mageddo.dnsproxyserver.docker.application.DockerConnectionCheck;
 import com.mageddo.dnsproxyserver.server.dns.solver.docker.Network;
-import com.mageddo.dnsproxyserver.server.dns.solver.docker.NetworkComparator;
+import com.mageddo.dnsproxyserver.server.dns.solver.docker.application.NetworkComparator;
 import com.mageddo.dnsproxyserver.server.dns.solver.docker.dataprovider.mapper.NetworkMapper;
 import com.mageddo.net.IP;
 import lombok.RequiredArgsConstructor;
